ml_app/requirements.txt
tensorflow==2.8.0
fastapi==0.78.0
python-multipart==0.0.5
pandas==1.4.3
scikit-learn==1.1.1
psycopg2-binary==2.9.3
hypercorn==0.13.2

protobuf==3.20.*
